Improve Alpine components code for CSP
