+++ assets/src/scripts/alpine/user-menu.js
@@ -1,0 +1,16 @@
+export default () => ({
+  isUserNavOpen: false,
+
+  showUserNav() {
+    this.isUserNavOpen = true;
+  },
+  hideUserNav() {
+    this.isUserNavOpen = false;
+  },
+
+  userNavAriaExpanded: {
+    [":aria-expanded"]() {
+      return this.isUserNavOpen.toString();
+    }
+  },
+});
