+++ templates/partials/header.html
@@ -26,7 +26,7 @@
                 class="inline-flex justify-center w-full rounded-md px-1 py-1 overflow-hidden bg-white text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-100 focus:ring-oxford-700"
                 id="menu-button"
                 type="button"
-                x-bind:aria-expanded="isUserNavOpen.toString()"
+                x-bind="userNavAriaExpanded"
                 x-on:click="showUserNav"
               >
                 <span class="sr-only">Options</span>
