+++ templates/partials/sidebar.html
@@ -70,23 +70,27 @@
           <nav class="flex-1 px-2 space-y-1 bg-oxford-800" aria-label="Sidebar">
             <ul>
               {% for category in categories %}
-              <li x-data="{ isSubmenuVisible: {% if report.category.id == category.id %}true{% else %}false{% endif %} }" class="space-y-1">
+              <li
+                {% if report.category.id == category.id %}data-active="true"{% endif %}"
+                class="space-y-1"
+                x-data="subnav"
+              >
                 <button
-                  @click="isSubmenuVisible = !isSubmenuVisible"
                   aria-controls="desktop-sidebar-{{ forloop.counter }}"
                   aria-expanded="false"
                   class="text-white hover:bg-oxford-900 hover:text-gray-50 group w-full flex items-center pr-2 py-2 text-left text-sm font-medium rounded-md focus:outline-none focus:ring-2 focus:ring-oxford-400"
                   type="button"
-                  x-bind:aria-expanded="isSubmenuVisible.toString()"
+                  x-bind="subnavAriaExpanded"
+                  x-on:click="toggleSubnav"
                 >
-                  {% include "icons/outline/chevron-right.svg" with htmlClass="text-gray-300 mr-2 flex-shrink-0 h-5 w-5 transform group-hover:text-gray-400 transition-colors ease-in-out duration-150" attrs=":class=\"{ 'text-gray-400 rotate-90': isSubmenuVisible, 'text-gray-300': !(isSubmenuVisible) }\"" %}
+                  {% include "icons/outline/chevron-right.svg" with htmlClass="text-gray-300 mr-2 flex-shrink-0 h-5 w-5 transform group-hover:text-gray-400 transition-colors ease-in-out duration-150" attrs="x-bind='subnavArrowStyles'" %}
                   {{ category.name }}
                 </button>
                 <ul
                   class="space-y-1"
                   id="desktop-sidebar-{{ forloop.counter }}"
-                  x-show="isSubmenuVisible"
                   x-cloak
+                  x-show="isSubmenuVisible"
                 >
                   {% category_reports_for_user category as category_reports %}
                   {% for single_report in category_reports %}
